refactored support for generic vm and qemu env vars
