+++ clab/register.go
@@ -14,6 +14,7 @@ import (
 	cvx "github.com/srl-labs/containerlab/nodes/cvx"
 	ext_container "github.com/srl-labs/containerlab/nodes/ext_container"
 	fortinet_fortigate "github.com/srl-labs/containerlab/nodes/fortinet_fortigate"
+	generic_vm "github.com/srl-labs/containerlab/nodes/generic_vm"
 	host "github.com/srl-labs/containerlab/nodes/host"
 	ipinfusion_ocnos "github.com/srl-labs/containerlab/nodes/ipinfusion_ocnos"
 	k8s_kind "github.com/srl-labs/containerlab/nodes/k8s_kind"
@@ -29,7 +30,6 @@ import (
 	vr_freebsd "github.com/srl-labs/containerlab/nodes/vr_freebsd"
 	vr_ftdv "github.com/srl-labs/containerlab/nodes/vr_ftdv"
 	vr_ftosv "github.com/srl-labs/containerlab/nodes/vr_ftosv"
-	vr_linux "github.com/srl-labs/containerlab/nodes/vr_linux"
 	vr_n9kv "github.com/srl-labs/containerlab/nodes/vr_n9kv"
 	vr_openbsd "github.com/srl-labs/containerlab/nodes/vr_openbsd"
 	vr_pan "github.com/srl-labs/containerlab/nodes/vr_pan"
@@ -66,7 +66,7 @@ func (c *CLab) RegisterNodes() {
 	vr_csr.Register(c.Reg)
 	vr_c8000v.Register(c.Reg)
 	vr_freebsd.Register(c.Reg)
-	vr_linux.Register(c.Reg)
+	generic_vm.Register(c.Reg)
 	vr_ftosv.Register(c.Reg)
 	vr_n9kv.Register(c.Reg)
 	vr_pan.Register(c.Reg)
