--- lab-examples/vr-linux01/vr-linux01.clab.yml
@@ -1,16 +1,0 @@
-name: vr-linux01
-
-topology:
-  nodes:
-    ceos:
-      kind: ceos
-      image: ceos:4.31.2F
-    ubuntu-01:
-      kind: vr_linux
-      image: vrnetlab/vr-ubuntu:jammy
-    ubuntu-02:
-      kind: vr_linux
-      image: vrnetlab/vr-ubuntu:jammy
-  links:
-    - endpoints: ["ceos:eth1", "ubuntu-01:eth1"]
-    - endpoints: ["ceos:eth2", "ubuntu-02:eth1"]
