+++ mkdocs.yml
@@ -42,7 +42,7 @@ nav:
           - KinD: manual/kinds/k8s-kind.md
           - Linux bridge: manual/kinds/bridge.md
           - Linux container: manual/kinds/linux.md
-          - Linux VM: manual/kinds/vr-linux.md
+          - Generic VM: manual/kinds/generic_vm.md
           - RARE/freeRtr: manual/kinds/rare-freertr.md
           - Openvswitch bridge: manual/kinds/ovs-bridge.md
           - External container: manual/kinds/ext-container.md
@@ -123,6 +123,7 @@ nav:
       - Templated labs:
           - Leaf-spine topology: lab-examples/templated01.md
           - 5-stage Clos topology: lab-examples/templated02.md
+      - Generic VM: lab-examples/generic_vm01.md
   - Release notes:
       - "0.54": rn/0.54.md
       - "0.53": rn/0.53.md
